requirements.txt
psd-tools
Pillow
numpy